+++ scripts/test/Ticket/Event/NotificationEvent.t
@@ -1019,6 +1019,25 @@ my @Tests = (
         Success => 1,
     },
     {
+        Name => 'RecipientCustomer + OncePerDay',
+        Data => {
+            Events               => [ 'TicketDynamicFieldUpdate_DFT1' . $RandomID . 'Update' ],
+            Recipients           => ['Customer'],
+            IsVisibleForCustomer => [0],
+            OncePerDay           => [1],
+        },
+        Config => {
+            Event => 'TicketDynamicFieldUpdate_DFT1' . $RandomID . 'Update',
+            Data  => {
+                TicketID => $TicketID,
+            },
+            Config => {},
+            UserID => 1,
+        },
+        ExpectedResults => [],
+        Success         => 1,
+    },
+    {
         Name => 'RecipientEmail filter by unchecked dynamic field',
         Data => {
             Events         => [ 'TicketDynamicFieldUpdate_DFT1' . $RandomID . 'Update' ],